+++ src/main/java/com/onevizion/uitest/api/helper/entity/EntityReportGroup.java
@@ -78,7 +78,7 @@ public class EntityReportGroup {
     }
 
     public void testInGrid(Long gridId, Long rowIndex, ReportGroup reportGroup) {
-        Map<Long, String> gridVals = new HashMap<>();
+        Map<Integer, String> gridVals = new HashMap<>();
 
         gridVals.put(js.getColumnIndexByLabel(gridId, "Report Group Name"), reportGroup.getName());
